+++ packages/site/src/pages/visualTests/VisualTestRouter.tsx
@@ -1,5 +1,6 @@
 import { Route, Switch } from "react-router";
 import { VisualTestAutocompletePage } from "./VisualTestAutocompletePage";
+import { VisualTestAutocompleteV2Page } from "./VisualTestAutocompleteV2Page";
 import { VisualTestChipPage } from "./VisualTestChipPage";
 import { VisualTestChipsPage } from "./VisualTestChipsPage";
 import { VisualTestComboboxPage } from "./VisualTestComboboxPage";
@@ -52,6 +53,7 @@ import { VisualTestTypographyPage } from "./VisualTestTypographyPage";
 import { VisualTestTabsPage } from "./VisualTestTabsPage";
 import { VisualTestToastPage } from "./VisualTestToastPage";
 import { VisualTestBannerPage } from "./VisualTestBannerPage";
+import { VisualTestThemePage } from "./VisualTestThemePage";
 
 export const VisualTestRouter = () => {
   return (
@@ -80,6 +82,11 @@ export const VisualTestRouter = () => {
         path="/visual-tests/autocomplete"
         exact={true}
         component={VisualTestAutocompletePage}
+      />
+      <Route
+        path="/visual-tests/autocomplete-v2"
+        exact={true}
+        component={VisualTestAutocompleteV2Page}
       />
       <Route
         path="/visual-tests/banner"
@@ -320,6 +327,11 @@ export const VisualTestRouter = () => {
         path="/visual-tests/typography"
         exact={true}
         component={VisualTestTypographyPage}
+      />
+      <Route
+        path="/visual-tests/theme"
+        exact={true}
+        component={VisualTestThemePage}
       />
     </Switch>
   );